refactor: rename LegendField to Fieldset with optional legend and attributes
The grouping field now models the fieldset itself: an optional \$legend (instead
of a mandatory label) and an HtmlAttributes bag applied to the <fieldset> element.
The field renders the opening <fieldset> tag (with attributes) plus the legend;
the handler still decides where to close it.
